Call /update instead of /install on the backend when updating apps
The badge has no way of knowing whether an installed app came from the
badge store or local development, so simply requests all installed apps
when doing an update. The default /install path on the backend will
return a 404 if any of the requested apps aren't in the repo.

This commit calls /update instead, which doesn't 404 if an app is
missing.

Depends on emfcamp/Mk4-Backend#2

Fixes #57